changed from decimal to float on the coin table

+++ db/migrate/20201014195358_create_coins.rb
@@ -3,9 +3,9 @@ class CreateCoins < ActiveRecord::Migration
     create_table :coins do |t|
       t.string :name
       t.string :symbol
-      t.integer :quantity
-      t.integer :amount_invested
-      t.integer :average_coin_price
+      t.decimal :quantity
+      t.float :amount_invested
+      t.float :average_coin_price
       t.string :user_id
       t.timestamps null: false
     end
